+++ src/Symfony/Component/Serializer/Normalizer/ContextAwareDenormalizerInterface.php
@@ -15,6 +15,8 @@ namespace Symfony\Component\Serializer\Normalizer;
  * Adds the support of an extra $context parameter for the supportsDenormalization method.
  *
  * @author Kévin Dunglas <dunglas@gmail.com>
+ *
+ * @deprecated since symfony/serializer 6.1, use DenormalizerInterface instead
  */
 interface ContextAwareDenormalizerInterface extends DenormalizerInterface
 {