+++ src/Symfony/Component/Serializer/Normalizer/NormalizerInterface.php
@@ -41,10 +41,11 @@ interface NormalizerInterface
     /**
      * Checks whether the given class is supported for normalization by this normalizer.
      *
-     * @param mixed  $data   Data to normalize
-     * @param string $format The format being (de-)serialized from or into
+     * @param mixed  $data    Data to normalize
+     * @param string $format  The format being (de-)serialized from or into
+     * @param array  $context Context options for the normalizer
      *
      * @return bool
      */
-    public function supportsNormalization(mixed $data, string $format = null);
+    public function supportsNormalization(mixed $data, string $format = null /*, array $context = [] */);
 }