+++ src/Symfony/Component/Serializer/Normalizer/ProblemNormalizer.php
@@ -64,8 +64,10 @@ class ProblemNormalizer implements NormalizerInterface, CacheableSupportsMethodI
 
     /**
      * {@inheritdoc}
+     *
+     * @param array $context
      */
-    public function supportsNormalization(mixed $data, string $format = null): bool
+    public function supportsNormalization(mixed $data, string $format = null /*, array $context = [] */): bool
     {
         return $data instanceof FlattenException;
     }